Move duplicated functions to artcommon
