+++ doozer/doozerlib/image.py
@@ -9,6 +9,7 @@ from copy import deepcopy
 
 import doozerlib
 from artcommonlib.arch_util import brew_arch_for_go_arch, go_arch_for_brew_arch
+from artcommonlib.release_util import isolate_el_version_in_release
 from doozerlib import brew, coverity, exectools
 from doozerlib.constants import BREWWEB_URL
 from doozerlib.distgit import pull_image
@@ -17,7 +18,6 @@ from doozerlib.model import Missing, Model
 from doozerlib.pushd import Dir
 from doozerlib.repodata import OutdatedRPMFinder, Repodata
 from doozerlib.rpm_utils import parse_nvr, to_nevra
-from doozerlib.util import isolate_el_version_in_release
 
 
 class ImageMetadata(Metadata):
