+++ elliott/elliottlib/util.py
@@ -258,50 +258,6 @@ def get_golang_version_from_build_log(log):
     return go_version
 
 
-def split_el_suffix_in_release(release: str) -> Tuple[str, Optional[str]]:
-    """
-    Given a release field, this will method will split out any
-    .el### suffix and return (prefix, el_suffix) where el_suffix
-    is None if there .el### is not detected.
-    """
-
-    el_suffix_match = re.match(r'(.*)\.(el\d+)(?:\.+|$)', release)
-    if el_suffix_match:
-        prefix = el_suffix_match.group(1)
-        el_suffix = el_suffix_match.group(2)
-        return prefix, el_suffix
-    else:
-        return release, None
-
-
-def isolate_assembly_in_release(release: str) -> str:
-    """
-    Given a release field, determines whether is contains
-    an assembly name. If it does, it returns the assembly
-    name. If it is not found, None is returned.
-    """
-    # Because RPM release fields will have .el? as their suffix, we cannot
-    # assume that endswith(.assembly.<name>). Strip off .el?
-    prefix, _ = split_el_suffix_in_release(release)
-    asm_pos = prefix.rfind('.assembly.')
-    if asm_pos == -1:
-        return None
-
-    return prefix[asm_pos + len('.assembly.'):]
-
-
-def isolate_el_version_in_release(release: str) -> Optional[int]:
-    """
-    Given a release field, determines whether is contains
-    a RHEL version. If it does, it returns the version value as an int.
-    If it is not found, None is returned.
-    """
-    _, el_suffix = split_el_suffix_in_release(release)
-    if el_suffix:
-        return int(el_suffix[2:])
-    return None
-
-
 def isolate_el_version_in_brew_tag(tag: str) -> Optional[int]:
     """
     Given a brew tag (target) name, determines whether is contains
@@ -310,44 +266,6 @@ def isolate_el_version_in_brew_tag(tag: str) -> Optional[int]:
     """
     el_version_match = re.search(r"rhel-(\d+)", tag)
     return int(el_version_match[1]) if el_version_match else None
-
-
-def find_latest_build(builds: List[Dict], assembly: Optional[str]) -> Optional[Dict]:
-    """ Find the latest build specific to the assembly in a list of builds belonging to the same component and brew tag
-    :param brew_builds: a list of build dicts sorted by tagging event in descending order
-    :param assembly: the name of assembly; None if assemblies support is disabled
-    :return: a brew build dict or None
-    """
-    chosen_build = None
-    if not assembly:  # if assembly is not enabled, choose the true latest tagged
-        chosen_build = builds[0] if builds else None
-    else:  # assembly is enabled
-        # find the newest build containing ".assembly.<assembly-name>" in its RELEASE field
-        chosen_build = next((build for build in builds if isolate_assembly_in_release(build["release"]) == assembly), None)
-        if not chosen_build and assembly != "stream":
-            # If no such build, fall back to the newest build containing ".assembly.stream"
-            chosen_build = next((build for build in builds if isolate_assembly_in_release(build["release"]) == "stream"), None)
-        if not chosen_build:
-            # If none of the builds have .assembly.stream in the RELEASE field, fall back to the latest build without .assembly in the RELEASE field
-            chosen_build = next((build for build in builds if isolate_assembly_in_release(build["release"]) is None), None)
-    return chosen_build
-
-
-def find_latest_builds(brew_builds: Iterable[Dict], assembly: Optional[str]) -> Iterable[Dict]:
-    """ Find latest builds specific to the assembly in a list of brew builds.
-    :param brew_builds: a list of build dicts sorted by tagging event in descending order
-    :param assembly: the name of assembly; None if assemblies support is disabled
-    :return: an iterator of latest brew build dicts
-    """
-    # group builds by component name
-    grouped_builds = {}  # key is component_name, value is a list of Brew build dicts
-    for build in brew_builds:
-        grouped_builds.setdefault(build["name"], []).append(build)
-
-    for builds in grouped_builds.values():  # builds are ordered from newest tagged to oldest tagged
-        chosen_build = find_latest_build(builds, assembly)
-        if chosen_build:
-            yield chosen_build
 
 
 def split_nvr_epoch(nvre):
