+++ elliott/tests/test_util.py
@@ -6,51 +6,6 @@ from elliottlib import brew
 
 
 class TestUtil(unittest.TestCase):
-    def test_isolate_assembly_in_release(self):
-        self.assertEqual(util.isolate_assembly_in_release('1.2.3-y.p.p1'), None)
-        self.assertEqual(util.isolate_assembly_in_release('1.2.3-y.p.p1.assembly'), None)
-        self.assertEqual(util.isolate_assembly_in_release('1.2.3-y.p.p1.assembly.x'), 'x')
-        self.assertEqual(util.isolate_assembly_in_release('1.2.3-y.p.p1.assembly.xyz'), 'xyz')
-        self.assertEqual(util.isolate_assembly_in_release('1.2.3-y.p.p1.assembly.xyz.el7'), 'xyz')
-        self.assertEqual(util.isolate_assembly_in_release('1.2.3-y.p.p1.assembly.4.9.99.el7'), '4.9.99')
-        self.assertEqual(util.isolate_assembly_in_release('1.2.3-y.p.p1.assembly.4.9.el700.hi'), '4.9')
-        self.assertEqual(util.isolate_assembly_in_release('1.2.3-y.p.p1.assembly.art12398.el10'), 'art12398')
-        self.assertEqual(util.isolate_assembly_in_release('1.2.3-y.p.p1.assembly.art12398.el10'), 'art12398')
-
-    def test_isolate_el_version_in_release(self):
-        self.assertEqual(util.isolate_el_version_in_release('1.2.3-y.p.p1.assembly.4.9.99.el7'), 7)
-        self.assertEqual(util.isolate_el_version_in_release('1.2.3-y.p.p1.assembly.4.9.el7'), 7)
-        self.assertEqual(util.isolate_el_version_in_release('1.2.3-y.p.p1.assembly.art12398.el199'), 199)
-        self.assertEqual(util.isolate_el_version_in_release('1.2.3-y.p.p1.assembly.art12398'), None)
-        self.assertEqual(util.isolate_el_version_in_release('1.2.3-y.p.p1.assembly.4.7.e.8'), None)
-
-    def test_find_latest_builds(self):
-        builds = [
-            {"id": 13, "name": "a-container", "version": "v1.2.3", "release": "3.assembly.stream", "tag_name": "tag1"},
-            {"id": 12, "name": "a-container", "version": "v1.2.3", "release": "2.assembly.hotfix_a", "tag_name": "tag1"},
-            {"id": 11, "name": "a-container", "version": "v1.2.3", "release": "1.assembly.hotfix_a", "tag_name": "tag1"},
-            {"id": 23, "name": "b-container", "version": "v1.2.3", "release": "3.assembly.test", "tag_name": "tag1"},
-            {"id": 22, "name": "b-container", "version": "v1.2.3", "release": "2.assembly.hotfix_b", "tag_name": "tag1"},
-            {"id": 21, "name": "b-container", "version": "v1.2.3", "release": "1.assembly.stream", "tag_name": "tag1"},
-            {"id": 33, "name": "c-container", "version": "v1.2.3", "release": "3", "tag_name": "tag1"},
-            {"id": 32, "name": "c-container", "version": "v1.2.3", "release": "2.assembly.hotfix_b", "tag_name": "tag1"},
-            {"id": 31, "name": "c-container", "version": "v1.2.3", "release": "1", "tag_name": "tag1"},
-        ]
-        actual = util.find_latest_builds(builds, "stream")
-        self.assertEqual([13, 21, 33], [b["id"] for b in actual])
-
-        actual = util.find_latest_builds(builds, "hotfix_a")
-        self.assertEqual([12, 21, 33], [b["id"] for b in actual])
-
-        actual = util.find_latest_builds(builds, "hotfix_b")
-        self.assertEqual([13, 22, 32], [b["id"] for b in actual])
-
-        actual = util.find_latest_builds(builds, "test")
-        self.assertEqual([13, 23, 33], [b["id"] for b in actual])
-
-        actual = util.find_latest_builds(builds, None)
-        self.assertEqual([13, 23, 33], [b["id"] for b in actual])
-
     def test_isolate_timestamp_in_release(self):
         actual = util.isolate_timestamp_in_release("foo-4.7.0-202107021813.p0.g01c9f3f.el8")
         expected = "202107021813"
